+++ Assets/Script/CommonLib/Requests/VerifyStageBattleRequest.cs
@@ -6,9 +6,9 @@ namespace Script.CommonLib.Requests
     {
         public ulong UserId { get; }
         public string RequestId { get; }
-        public List<float> UpdateIntervals { get; }
+        public List<ushort> UpdateIntervals { get; }
 
-        public VerifyStageBattleRequest(ulong userId, string requestId, List<float> updateIntervals)
+        public VerifyStageBattleRequest(ulong userId, string requestId, List<ushort> updateIntervals)
         {
             UserId = userId;
             RequestId = requestId;
